--- comandaproject/src/main/java/com/ghartmann/domain/ItemPedido.java
@@ -1,78 +1,0 @@
-package com.ghartmann.domain;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.SequenceGenerator;
-import javax.persistence.Table;
-
-@Entity
-@Table(name="ITEMPEDIDO_TABLE")
-public class ItemPedido {
-
-    @Id
-    @GeneratedValue(strategy=GenerationType.SEQUENCE, generator="id_pratomesa_sequence")
-    @SequenceGenerator(name="id_pratomesa_sequence", sequenceName="pratomesa_seq")
-    int id;
-
-    @Column(name="Observação")
-    String observação;
-
-    @ManyToOne
-    @JoinColumn(name = "pedido_id")
-    private Pedido pedido;
-
-    @ManyToOne
-    @JoinColumn(name = "prato_id")
-    private Prato prato;
-
-    @ManyToOne
-    @JoinColumn(name="mesa_id")
-    private Mesa mesa;
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public Mesa getMesa() {
-        return mesa;
-    }
-
-    public void setMesa(Mesa mesa) {
-        this.mesa = mesa;
-    }
-
-    public String getObservação() {
-        return observação;
-    }
-
-    public void setObservação(String observação) {
-        this.observação = observação;
-    }
-
-    public Pedido getPedido() {
-        return pedido;
-    }
-
-    public void setPedido(Pedido pedido) {
-        this.pedido = pedido;
-    }
-
-    public Prato getPrato() {
-        return prato;
-    }
-
-    public void setPrato(Prato prato) {
-        this.prato = prato;
-    }
-
-
-}
